requirements.txt
fastapi=0.116.1
uvicorn=0.35.0
qlmodel=0.0.24
argon2_cffi=25.1.0
PyJWT=2.10.1
python_dotenv=1.1.1
pymysql=1.1.2